+++ apps/restconf/restconf_main_evhtp.c
@@ -80,7 +80,7 @@
 #include "restconf_root.h"
 
 /* Command line options to be passed to getopt(3) */
-#define RESTCONF_OPTS "hD:f:l:p:d:y:a:u:ro:scP:"
+#define RESTCONF_OPTS "hD:f:E:l:p:d:y:a:u:ro:scP:"
 
 /* See see listen(5) */
 #define SOCKET_LISTEN_BACKLOG 16
@@ -594,6 +594,7 @@ usage(clicon_handle h,
             "\t-h \t\t  Help\n"
 	    "\t-D <level>\t  Debug level\n"
     	    "\t-f <file>\t  Configuration file (mandatory)\n"
+	    "\t-E <dir> \t  Extra configuration file directory\n"
 	    "\t-l <s|f<file>> \t  Log on (s)yslog, (f)ile (syslog is default)\n"
 	    "\t-p <dir>\t  Yang directory path (see CLICON_YANG_DIR)\n"
 	    "\t-d <dir>\t  Specify restconf plugin directory dir (default: %s)\n"
@@ -666,6 +667,11 @@ main(int    argc,
 		usage(h, argv0);
 	    clicon_option_str_set(h, "CLICON_CONFIGFILE", optarg);
 	    break;
+	case 'E': /* extra config directory */
+	    if (!strlen(optarg))
+		usage(h, argv[0]);
+	    clicon_option_str_set(h, "CLICON_CONFIGDIR", optarg);
+	    break;
 	 case 'l': /* Log destination: s|e|o */
 	     if ((logdst = clicon_log_opt(optarg[0])) < 0)
 		usage(h, argv0);
@@ -715,6 +721,7 @@ main(int    argc,
 	case 'h' : /* help */
 	case 'D' : /* debug */
 	case 'f':  /* config file */
+	case 'E':  /* extra config dir */
 	case 'l':  /* log  */
 	    break; /* see above */
 	case 'p' : /* yang dir path */